use nestjs monorepo to separate out NestJS server and worker process
