+++ nestjs-counter/apps/server/src/app.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { CounterWorkflowModule } from './counter-workflow/counter-workflow.module';
+@Module({
+  imports: [CounterWorkflowModule],
+  controllers: [],
+  providers: [],
+})
+export class AppModule {}
