+++ nestjs-counter/apps/server/test/jest-e2e.json
