+++ nestjs-counter/apps/worker/src/counter-worker/counter-worker.providers.ts
@@ -1,6 +1,6 @@
 import { ActivitiesService } from '../activities/activities.service';
 import { DefaultLogger, LogEntry, Runtime, Worker } from '@temporalio/worker';
-import { taskQueue } from '../temporal/shared';
+import { taskQueue } from '@app/shared';
 import { ActivityInterface } from '@temporalio/workflow';
 
 export const counterWorkerProviders = [
