+++ nestjs-counter/apps/worker/src/temporal/workflows.ts
@@ -1,14 +1,6 @@
-import {
-  condition,
-  defineQuery,
-  defineSignal,
-  proxyActivities,
-  setHandler,
-} from '@temporalio/workflow';
+import { condition, proxyActivities, setHandler } from '@temporalio/workflow';
 import { IActivities } from './activities';
-
-export const incrementSignal = defineSignal<[number]>('increment');
-export const getValueQuery = defineQuery<number>('getValue');
+import { incrementSignal, getValueQuery } from '@app/shared';
 
 const { persist } = proxyActivities<IActivities>({
   startToCloseTimeout: '30 seconds',
@@ -18,7 +10,6 @@ export async function counterWorkflow(initial: number): Promise<void> {
   let value = initial;
 
   setHandler(incrementSignal, async (val: number) => {
-    console.log('Increment', val);
     if (!val) {
       return;
     }
