+++ nestjs-counter/package.json
@@ -7,19 +7,13 @@
   "scripts": {
     "prebuild": "rimraf dist",
     "build": "nest build",
-    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "start": "nest start",
+    "format": "prettier --write \"apps/**/*.ts\" \"libs/**/*.ts\"",
+    "start": "nest start server",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "test": "env NODE_OPTIONS=--experimental-vm-modules jest",
-    "test:watch": "env NODE_OPTIONS=--experimental-vm-modules jest --watch",
-    "test:cov": "env NODE_OPTIONS=--experimental-vm-modules jest --coverage",
-    "test:debug": "env NODE_OPTIONS=--experimental-vm-modules node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "env NODE_OPTIONS=--experimental-vm-modules jest --config ./test/jest-e2e.json",
-    "worker": "ts-node src/worker.ts",
-    "worker:watch": "nodemon src/worker.ts"
+    "worker": "nest start worker"
   },
   "dependencies": {
     "@nestjs/common": "^8.1.1",
@@ -62,7 +56,7 @@
       "json",
       "ts"
     ],
-    "rootDir": "src",
+    "rootDir": ".",
     "testRegex": ".*\\.spec\\.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
@@ -70,7 +64,14 @@
     "collectCoverageFrom": [
       "**/*.(t|j)s"
     ],
-    "coverageDirectory": "../coverage",
-    "testEnvironment": "node"
+    "coverageDirectory": "./coverage",
+    "testEnvironment": "node",
+    "roots": [
+      "<rootDir>/apps/",
+      "<rootDir>/libs/"
+    ],
+    "moduleNameMapper": {
+      "^@app/shared(|/.*)$": "<rootDir>/libs/shared/src/$1"
+    }
   }
 }
