--- nestjs-counter/src/temporal/shared.ts
@@ -1,1 +1,0 @@
-export const taskQueue = 'nest-test';
